+++ data/schema.graphql
@@ -230,8 +230,8 @@ type RootMutationType {
   """Creates repositories"""
   createRepositories(admins: [String!], arePrivate: Boolean, courseId: ID!, maintainers: [String!], names: [String!], organization: String!, repositoriesData: [RepositoryStudentData!]): CreateRepositoriesResponse
 
-  """Create a review"""
-  createReview(courseId: ID!, grade: Int, revisionRequested: Boolean!, submissionId: ID!): InternalReviewType!
+  """Create a review within a submission"""
+  createReview(courseId: ID!, grade: Int, revisionRequested: Boolean!, submissionId: ID!): SubmissionType!
 
   """Creates a new submission for the viewer"""
   createSubmission(assignmentId: ID!, courseId: ID!, description: String, pullRequestUrl: String!): SubmissionType!
